♻️ Refactor deny to repository and with category
