+++ src/lib/server/types/validation/submissions.ts
@@ -6,3 +6,11 @@ export const approveRequestSchema = object({
 	modNotes: string().nullable().max(500)
 });
 export type ApproveRequest = InferType<typeof approveRequestSchema>;
+
+export const denyRequestSchema = object({
+	moderatorName: string().required(),
+	runId: number().required(),
+	modNotes: string().nullable().max(500)
+});
+
+export type DenyRequest = InferType<typeof denyRequestSchema>;
