+++ src/routes/ngs-api/submissions/[category=submissionCategory]/deny/+server.ts
@@ -1,0 +1,68 @@
+import { leaderboardDb } from '$lib/server/db/db';
+import { json } from '@sveltejs/kit';
+import { jsonError } from '$lib/server/error.js';
+import {
+	denyIndomitableSubmission,
+	getIndomitableExists
+} from '$lib/server/repositories/indomitableSubmissionsRepository.js';
+import { denyRequestSchema, type DenyRequest } from '$lib/server/types/validation/submissions.js';
+
+/** @type {import('./$types').RequestHandler} */
+export async function POST({ request, params }) {
+	const category = params.category?.toLowerCase() ?? '';
+
+	// Validate request
+	const body = await request.json();
+	let submission: DenyRequest;
+	try {
+		submission = await denyRequestSchema.validate(body);
+	} catch (err: any) {
+		return jsonError(400, {
+			error: 'bad_request',
+			details: err.errors
+		});
+	}
+	if (!submission) {
+		return jsonError(400, { error: 'json_parse_error' });
+	}
+
+	// Check data in db
+	const { errorList: validationErrors, extra: data } = await checkData(submission, category);
+	if (validationErrors.length > 0 || !data) {
+		return jsonError(400, { error: 'bad_request', details: validationErrors });
+	}
+
+	try {
+		const pool = await leaderboardDb.connect();
+		await denyIndomitableSubmission(pool.request(), category, submission);
+		return json({ data: 'success' });
+	} catch (err) {
+		console.error(err);
+		throw jsonError(500, { error: 'internal_server_error' });
+	}
+}
+
+const checkData = async (run: DenyRequest, category: string) => {
+	const pool = await leaderboardDb.connect();
+	const errorList: string[] = [];
+
+	// Run exists
+	const submissionRequest = pool.request();
+	const submissionResult = await getIndomitableExists(submissionRequest, category, run.runId);
+	if (!submissionResult) {
+		errorList.push(`Unknown submissionId`);
+		return {
+			errorList
+		};
+	}
+	if (submissionResult && submissionResult.SubmissionStatus != 0) {
+		errorList.push(`Submission already denied/approved`);
+	}
+
+	return {
+		errorList,
+		extra: {
+			playerId: parseInt(submissionResult.PlayerId)
+		}
+	};
+};
